// src/app/shared/models/basic-info.model.ts
export class BasicInfo {
  registration_type: number;
  purpose_of_insurance: number;
  //  sequence_number: string;
  //  custom_number: string;
  insured_national_id: string;
  policy_effective_date: any;
  birthMonth?: string;
  birthYear?: string;
  manifucture_year: any;
  VehicleID: number;
}